# Fix Inefficient String Encoding