The code for constructing `List(v::Vector{String})` is extremely slow, causing problems when writing Feather files of dataframes that are string heavy (for instance, it took me 770s to write a 3.2GB file with a large number of address fields, whereas python can write the same file in 12s).  It looks like the problem comes from the code that [encodes](https://github.com/ExpandingMan/Arrow.jl/blob/94b335e209485b15163af974e6cb58bd34632e76/src/utils.jl#L63)  the strings as vectors of `UInt8`.  

I played around with it, and it looks like you can get nearly a 20X speedup by writing a specialized method to handle vectors of strings as in this PR:
```julia
julia> using Arrow, BenchmarkTools
julia> x = repeat(["a", "b", "c"].^10, 5_000_000);
# Before this PR
julia> @btime arrowformat($x)
  8.794 s (45016393 allocations: 69.95 GiB)
# After this PR
julia> @btime arrowformat($x)
  463.789 ms (13 allocations: 400.54 MiB)
```
I'm not super familiar with the internals of string encodings, but this seems to work on my local machine (and passes all the tests).  I've tried it out with Feather as well on some real world datasets, and everything seems to work properly.