@ExpandingMan 2019-03-24
That's extremely helpful, thanks!  I must confess, when I wrote this I was much more concerned with *deserialization* than serialization, so I did not spend that much time benchmarking the serialization, perhaps not at all for strings (which is pretty much the only non-trivial case that this package now supports).

I have no idea why stupid travis is not triggering.  I *always* have problems with it.  It never triggers and I don't know why, it's extremely annoying.  So, I have to try to figure that out before merging.

Also, be advised that I am in the process of doing a major re-write of this package to support the full standard and IPC metadata.  There may be some regressions by the time I'm finished, but I will probably merge that into master quite a while before I tag it, so there will be ample opportunity to fix those.  Things will become considerably trickier, because now I must support much more general cases, for example the underlying `List` object must be appropriate for both `String` and `Vector` (though `List` will work somewhat differently, as the underlying `List` will only return `Vector` objects from `getindex` operations, and there will be a wrapper `AbstractVector` which will have a `getindex` method which returns `String`s).

Anyway, if I get too frustrated with travis, I will test this locally and then merge.  Thanks.