@jacobadenbaum 2019-03-24
It looks like Travis did get [triggered](https://travis-ci.org/ExpandingMan/Arrow.jl/builds/510634201), but it just isn't hooking into the PR