@ExpandingMan 2019-03-24
I see no reason to wait to merge this, so I'll do so now.  Thanks again.